ConciliaciónBancariaEscritorio/requirements.txt
psycopg2-binary
pdfplumber>=0.11.0
pandas
matplotlib
tkcalendar
